ssim: Replace unsigned long with uint32_t.
The assembly only writes the low 4 bytes, and the HBD version only uses
uint32_t bytes.

+++ vpx_dsp/ssim.h
@@ -29,19 +29,19 @@ void vpx_reset_mmx_state(void);
 // metrics used for calculating ssim, ssim2, dssim, and ssimc
 typedef struct {
   // source sum ( over 8x8 region )
-  uint64_t sum_s;
+  uint32_t sum_s;
 
   // reference sum (over 8x8 region )
-  uint64_t sum_r;
+  uint32_t sum_r;
 
   // source sum squared ( over 8x8 region )
-  uint64_t sum_sq_s;
+  uint32_t sum_sq_s;
 
   // reference sum squared (over 8x8 region )
-  uint64_t sum_sq_r;
+  uint32_t sum_sq_r;
 
   // sum of source times reference (over 8x8 region)
-  uint64_t sum_sxr;
+  uint32_t sum_sxr;
 
   // calculated ssim score between source and reference
   double ssim;
